KB-12-2 rename
