+++ src/main/kotlin/lesson_12/lesson_12_2.kt
@@ -1,6 +1,15 @@
-package lesson_12_2
+package lesson_12
 
-class KeepWeather(_temperatureDay: Int, _temperatureNight: Int, _rain: Boolean, _atmosphericPressure: Int) {
+fun main(){
+    val thirdDayLife = KeepWeatherTask2(12, 10, true, 660)
+    val fourthDayLife = KeepWeatherTask2(0, -50, false, 550)
+    val fifthDayLife = KeepWeatherTask2(0, 50, true, 550)
+
+    thirdDayLife.printWeather()
+    fourthDayLife.printWeather()
+    fifthDayLife.printWeather()
+}
+class KeepWeatherTask2(_temperatureDay: Int, _temperatureNight: Int, _rain: Boolean, _atmosphericPressure: Int) {
 
     var temperatureDay = _temperatureDay
     var temperatureNight = _temperatureNight
